[Justice Counts] Handle Sub-Dimension without Data when converting Disaggregations to JSON (Recidiviz/recidiviz-data#41404)
## Description of the change

A
[KeyError](https://recidiviz-inc.sentry.io/issues/6562943368/?project=4504532096516096&query=is%3Aunresolved%20issue.priority%3A%5Bhigh%2C%20medium%5D&referrer=issue-stream&stream_index=1)
was surfaced in Sentry when Ilya was playing with the BE for
sub-dimensions. I added a check that the dimension is actually in
`dimension_to_other_sub_dimension_to_value` before trying to fetch the
value.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 74e97b28c81073bfa7d65b0453fc99ceb64752cd