[Sentencing] Validation for case_insights_rates (Recidiviz/recidiviz-data#41421)
## Description of the change

> Validation to check that every charge description in
`normalized_state.state_charge_v2` is present in `case_insights_rates`.
A validation view, `case_insights_rates_missing_charges` was created and
validated in a sandbox. The validation query currently only selects
charges in ND and IX to check against, as those are the only states for
which `case_insights_rates` currently has data for. Hard override of 5
rows of missing charges, with a soft override of 0 rows, as we want to
know any instance of charges not making it into `case_insights_rates`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#36042

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a7329c41d4f979e5f05f1987344493e94f2bdb93