[Security] Use Gemini for PII check (Recidiviz/recidiviz-data#40794)
> We currently use Nightfall. Nightfall can be error-prone and includes
licensing cost. This PR is an attempt to offload our PII checks to
Gemini (we are using a model in Vertex AI Studio in the
`recidiviz-security` project)

AI disclosure, the base script for this GH action was generated with
Vertex AI and then modified on top of that.

## Type of change

This pull request introduces a new script and workflow to detect
Personally Identifiable Information (PII) in pull request changes using
a generative AI model. Additionally, it includes a new data file for
testing purposes. The most important changes are summarized below:

### New PII Detection Script and Workflow:

* **PII Detection Script**: Added a script `pii_check.py` that analyzes
code changes in a pull request for potential PII using a generative AI
model. The script updates the pull request with its findings.

* **GitHub Workflow for PII Detection**: Created a new GitHub Actions
workflow `pii-check.yml` to run the PII detection script on pull request
events. The workflow sets up the necessary environment, checks out the
code, installs dependencies, and runs the PII analyzer.

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: a50e70131abffd83cf455400d2e034c02dd47ba9