[US_AZ] Define google sheet backed table for ORAS data (Recidiviz/recidiviz-data#41468)
## Description of the change

This is a first step to importing data from ORAS to our raw data. Adds a
google sheet backed table definition and a resource block to
`bigquery.tf` that will make a table that uses the google sheet as a
source. I think that since the entire `google_sheet_backed_tables`
module is included in `bigquery-scheduled-queries.tf`, and this new
sheet is configured within that module, it will automatically be
refreshed each day.

The default service accounts for `recidiviz-staging` and `recidiviz-123`
have view access to the relevant Google Sheet where this data will come
from.

One thing I am wondering about: the google sheet has terrible field
names. They have all sorts of disallowed characters. The tf config for
the google-backed table says to skip the first row, and defines a
schema; the block in `bigquery.tf` that makes the table defines the
schema according to that config, and also skips the first row. I defined
the field names I _want_ in the config - will it matter if the field
names in the file do not match that?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 77e0fe3ab78c16687b7cf17b02ebf69276e92021