[US_IA][Workflows] Pull sex offender specialties from case type instead of raw text (Recidiviz/recidiviz-data#41265)
## Description of the change

This PR changes the IA state-specific criterion that checks for sex
offender specialties to a state-agnostic criterion that checks for sex
offense case types for the early discharge opportunity.

Ran a sandbox to check eligibility a few days ago and this resulted in 9
less people being eligible for early discharge overall, but I confirmed
with Brenda that this was due to a lag between raw data and ingested
data rather than the logic itself being different.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0dd70daafa524c08cf20b1eada3807fac2adb4a2