[US_AZ] Change join logic in incarceration incidents view (Recidiviz/recidiviz-data#41517)
## Description of the change

There are three main tables connected to gather information about an
incarceration incident and its resulting outcomes: review, penalty, and
sanction. There can be many penalties per review, and one sanction per
penalty. We were previously joining sanction using the PK from the
review table instead of the PK from the sanction table, so we were
joining every sanction for a given review to every penalty for that
review. This PR fixes that join so that we are only joining the sanction
that is relevant _to a particular penalty_ to each penalty.

[sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-04-25_06_50_40-13081228501315353563?project=recidiviz-staging)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b84c03ba5a6323654d0033e59c7f0ccc0fd4c600