update outliers tests to use US_XX state code and a mock backend config (Recidiviz/recidiviz-data#41493)
## Description of the change

Right now, our tests + fixture data are tied to the US_PA state code.
This is fine, but it means unless we mock the backend config in a test,
we're going to be using the default US_PA config, which might not be
what we want. This makes it a bit easier to test upcoming changes that
affect non-PA states by making the tests more generic.

Inspired by this comment:
https://github.com/Recidiviz/recidiviz-data/pull/40872#issuecomment-2828496512

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 47915a966932a7df6e231d6fa806160fd5520d2a