[bigquery][airflow] add task to apply row level permissions (Recidiviz/recidiviz-data#41474)
## Description of the change

Currently, we apply row level permissions on table creation or view
materialization. however, the operation is very flaky and has been
causing our view update to fail which is expensive to rerun.

for now, i left the row level permission application on table
create/view materialize but don't fail the operation if row level
permissions fails. i've also added a task at the end of the calc dag to
apply permissions to any tables that do not have the correct policy in
place.

i didn't remove the row level permission application on table
create/view materialize because i have very little faith in this task to
apply permissions to a large number of tables without failing. i think
we will need to make changes to this setup for stability/performance
reasons but i just wanted to get a change out bc i'm about to be ooto
and i dont want us to keep failing the view update.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#41470

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9c160dcc32dde53c913338de3662d28ab18ad0fa