Don't comment if no PII detected (Recidiviz/recidiviz-data#41613)
## Description of the change

> It is noisy to hear everytime that PII is not detected. Disabling
comments if there are no findings

Copilot-generated comment

This pull request updates the `pii_check.py` script to conditionally
post comments on pull requests based on whether potential PII is
detected. The changes involve modifying the `update_pr_comment` function
to include a new `findings` parameter and updating its usage throughout
the codebase.

### Conditional Comment Posting:

*
[`recidiviz/tools/pii_detection/pii_check.py`](diffhunk://#diff-4fc8e041d74e09940af6bcfa141cb523dcd3e3d605a9c22a4fee91586e89b089L133-R133):
Added a `findings` parameter to the `update_pr_comment` function to
determine whether a comment should be posted. The function now checks
this parameter before making a POST request to add a comment.
[[1]](diffhunk://#diff-4fc8e041d74e09940af6bcfa141cb523dcd3e3d605a9c22a4fee91586e89b089L133-R133)
[[2]](diffhunk://#diff-4fc8e041d74e09940af6bcfa141cb523dcd3e3d605a9c22a4fee91586e89b089R156)

### Integration with Main Logic:

*
[`recidiviz/tools/pii_detection/pii_check.py`](diffhunk://#diff-4fc8e041d74e09940af6bcfa141cb523dcd3e3d605a9c22a4fee91586e89b089R223):
Updated calls to `update_pr_comment` in the `main` function to pass
`True` or `False` for the `findings` parameter, depending on whether
potential PII is detected. This ensures comments are only posted when
necessary.
[[1]](diffhunk://#diff-4fc8e041d74e09940af6bcfa141cb523dcd3e3d605a9c22a4fee91586e89b089R223)
[[2]](diffhunk://#diff-4fc8e041d74e09940af6bcfa141cb523dcd3e3d605a9c22a4fee91586e89b089L238-R240)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c41ff60f39a6e108806e3be2c38a7d4f6596f4ce