[Insights][US_MI] Add `feature_variant` and `inverse_feature_variant` options to OutliersMetricConfig (Recidiviz/recidiviz-data#40872)
## Description of the change

Add `feature_variant` and `inverse_feature_variant` options to
OutliersMetricConfig. This will let us configure the new
`splitProbationParoleOutcomes` FV on the new metrics and put it as an
`inverse_feature_variant` on the existing metrics so that users only see
one or the other. Update the querier to filter outcomes metrics
appropriately.

Adds a helper function that takes the user's FVs and returns the
relevant outcomes metrics from the outlier config so that we can filter
the DB for the relevant metrics.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues
 Recidiviz/recidiviz-data#39162
Closes Recidiviz/recidiviz-data#40360

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Dana Hoffman <1130694+danawillow@users.noreply.github.com>
GitOrigin-RevId: 4edc32a589361471de76816ac15239da2f0cc098